# api-gateway/docs/swagger.yaml
...
openapi: 3.0.3
info:
  title: Granula API
  version: 1.0.0
  description: |
    # Granula API

    API для интеллектуального сервиса планирования ремонта и перепланировки квартир.

    ## Возможности

    - 🏠 **Загрузка планировок** - Загружайте изображения планировок для автоматического распознавания
    - 🤖 **AI распознавание** - Автоматическое распознавание комнат, стен и элементов
    - 🎨 **3D редактор** - Редактирование планировки в интерактивном 3D редакторе
    - ✅ **Проверка норм** - Автоматическая проверка на соответствие СНиП и ЖК РФ
    - 💡 **AI генерация** - Генерация вариантов планировки с помощью AI
    - 📋 **Заявки экспертам** - Подача заявок на консультации специалистов БТИ

    ## Аутентификация

    API использует JWT (JSON Web Tokens) для аутентификации.

    ### Получение токенов

    1. Зарегистрируйтесь через `POST /api/v1/auth/register`
    2. Или войдите через `POST /api/v1/auth/login`
    3. Получите `access_token` и `refresh_token`

    ### Использование токенов

    Передавайте Access Token в заголовке Authorization:
    ```
    Authorization: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
    ```

    ### Обновление токенов

    Access Token действует 15 минут. Для обновления используйте `POST /api/v1/auth/refresh`

    ## Rate Limiting

    | Тип запроса | Лимит |
    |-------------|-------|
    | Анонимные | 60 req/min |
    | Аутентифицированные | 300 req/min |
    | AI endpoints | 30 req/min |

    При превышении лимита возвращается HTTP 429.

...
servers:
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT токен аутентификации. Получите через `/auth/login` или `/auth/register`.

        Формат: `Bearer <token>`

  # ===========================================================================
  # Schemas
  # ===========================================================================
  schemas:
...
    Pagination:
      type: object
      properties:
        total:
          type: integer
          description: Общее количество элементов
          example: 150
        page:
          type: integer
          description: Текущая страница
          example: 1
        page_size:
          type: integer
          description: Элементов на странице
          example: 20
        total_pages:
          type: integer
          description: Всего страниц
          example: 8
        has_next:
          type: boolean
          description: Есть следующая страница
          example: true
        has_prev:
          type: boolean
          description: Есть предыдущая страница
          example: false

    # -------------------------------------------------------------------------
    # Auth Schemas
    # -------------------------------------------------------------------------
...
    ExpertRequest:
      type: object
      properties:
        id:
          type: string
          format: uuid
        workspace_id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        title:
          type: string
          example: Консультация по перепланировке
        description:
          type: string
        category:
          type: string
          enum: [consultation, documentation, expert_visit, full_package]
          description: |
            Категория услуги:
            - **consultation** - Консультация (2000₽)
            - **documentation** - Подготовка документации (15000₽)
            - **expert_visit** - Выезд эксперта (5000₽)
            - **full_package** - Полный пакет (30000₽)
        priority:
          type: string
          enum: [low, normal, high, urgent]
          example: normal
        status:
          type: string
          enum: [draft, pending, in_review, approved, rejected, assigned, in_progress, completed, cancelled]
          description: |
            Статус заявки:
            - **draft** - Черновик
            - **pending** - Отправлена на рассмотрение
            - **in_review** - На рассмотрении
            - **approved** - Одобрена
            - **rejected** - Отклонена
            - **assigned** - Назначен эксперт
            - **in_progress** - В работе
            - **completed** - Завершена
            - **cancelled** - Отменена
          example: pending
        expert_id:
          type: string
          format: uuid
        assigned_at:
          type: string
          format: date-time
        estimated_cost:
          type: integer
          description: Оценочная стоимость (рубли)
          example: 2000
        final_cost:
          type: integer
          description: Финальная стоимость (рубли)
        rejection_reason:
          type: string
        contact_phone:
          type: string
          example: "+79991234567"
        contact_email:
          type: string
          format: email
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
    CreateExpertRequestRequest:
      type: object
      required:
        - workspace_id
        - title
        - description
        - category
      properties:
        workspace_id:
          type: string
          format: uuid
        title:
          type: string
          minLength: 5
          maxLength: 200
          example: Консультация по перепланировке 3-комнатной квартиры
        description:
          type: string
          minLength: 10
          maxLength: 5000
        category:
          type: string
          enum: [consultation, documentation, expert_visit, full_package]
        priority:
          type: string
          enum: [low, normal, high, urgent]
          default: normal
        contact_phone:
          type: string
        contact_email:
          type: string
          format: email

    # -------------------------------------------------------------------------
    # AI Schemas
    # -------------------------------------------------------------------------
...
paths:
...
  /requests:
    get:
      tags: [requests]
      summary: Список заявок
      operationId: listRequests
      security:
        - BearerAuth: []
      parameters:
        - name: workspace_id
          in: query
          schema:
            type: string
            format: uuid
        - name: status
          in: query
          schema:
            type: string
            enum: [draft, pending, in_review, approved, rejected, assigned, in_progress, completed, cancelled]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: page_size
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Список заявок
          content:
            application/json:
              schema:
                type: object
                properties:
                  requests:
                    type: array
                    items:
                      $ref: '#/components/schemas/ExpertRequest'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
    post:
      tags: [requests]
      summary: Создать заявку
      operationId: createRequest
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateExpertRequestRequest'
      responses:
        '201':
          description: Заявка создана
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExpertRequest'
  /requests/{id}/submit:
    post:
      tags: [requests]
      summary: Отправить заявку
      description: Переводит заявку из draft в pending
      operationId: submitRequest
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Заявка отправлена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExpertRequest'
...